pyproject.toml
[tool.poetry]
name = "flyps"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest-watcher = "^0.3.4"
pytest = "^7.4.4"
uvicorn = "^0.26.0"
loguru = "^0.7.2"
pendulum = "^3.0.0"
strawberry-graphql = {extras = ["debug-server"], version = "^0.217.1"}
rich = "^13.7.0"
typer = "^0.9.0"
poethepoet = "^0.24.4"
redis = "^5.0.1"
sqlalchemy = "^2.0.25"
zstd = "^1.5.5.1"
fastapi = "^0.109.0"
psycopg2-binary = "^2.9.9"

[tool.poe.tasks]
pytest = "ptw --now --verbose --exitfirst ."
server = "strawberry server flyps.graphql"
export = "strawberry export-schema flyps.graphql:schema --output schema.graphql"
psql15 = "podman run --rm -ti --env-file .env --name psql --publish 5432:5432 docker.io/library/postgres:15-alpine"

[tool.pycodestyle]
ignore = "E741"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
